Package rpc: Split code into multiple files.
